refactor: reorganize project src layout
