+++ src/assets/icons/icon128.png
