+++ src/assets/images/eth-qr.png
